+++ src/sdr/_modulation/_pulse_shapes.py
@@ -82,9 +82,7 @@ def raised_cosine(alpha: float, span: int, sps: int) -> np.ndarray:
     if not span * sps % 2 == 0:
         raise ValueError("The order of the filter (span * sps) must be even.")
 
-    # NOTE: Do the math in float64 to avoid numerical issues and then convert to float32 at the end
-
-    t = np.arange(-(span * sps) // 2, (span * sps) // 2 + 1, dtype=np.float64)
+    t = np.arange(-(span * sps) // 2, (span * sps) // 2 + 1, dtype=float)
     Ts = sps
 
     # Handle special cases where the denominator is zero
@@ -102,7 +100,7 @@ def raised_cosine(alpha: float, span: int, sps: int) -> np.ndarray:
     # Make the filter have unit energy
     h /= np.sqrt(np.sum(np.abs(h) ** 2))
 
-    return h.astype(np.float32)
+    return h
 
 
 @export
@@ -179,9 +177,7 @@ def root_raised_cosine(alpha: float, span: int, sps: int) -> np.ndarray:
     if not span * sps % 2 == 0:
         raise ValueError("The order of the filter (span * sps) must be even.")
 
-    # NOTE: Do the math in float64 to avoid numerical issues and then convert to float32 at the end
-
-    t = np.arange(-(span * sps) // 2, (span * sps) // 2 + 1, dtype=np.float64)
+    t = np.arange(-(span * sps) // 2, (span * sps) // 2 + 1, dtype=float)
     Ts = sps  # Symbol duration (in samples)
 
     # Handle special cases where the denominator is zero
@@ -201,7 +197,7 @@ def root_raised_cosine(alpha: float, span: int, sps: int) -> np.ndarray:
     # Make the filter have unit energy
     h /= np.sqrt(np.sum(np.abs(h) ** 2))
 
-    return h.astype(np.float32)
+    return h
 
 
 @export
@@ -265,9 +261,7 @@ def gaussian(time_bandwidth: float, span: int, sps: int) -> np.ndarray:
     if not span * sps % 2 == 0:
         raise ValueError("The order of the filter (span * sps) must be even.")
 
-    # NOTE: Do the math in float64 to avoid numerical issues and then convert to float32 at the end
-
-    t = np.arange(-(span * sps) // 2, (span * sps) // 2 + 1, dtype=np.float64)
+    t = np.arange(-(span * sps) // 2, (span * sps) // 2 + 1, dtype=float)
     t /= sps
 
     # Equation B.2
@@ -280,4 +274,4 @@ def gaussian(time_bandwidth: float, span: int, sps: int) -> np.ndarray:
     # Normalize coefficients so passband gain is 1
     h = h / np.sum(h)
 
-    return h.astype(np.float32)
+    return h
