+++ tests/dsp/test_to_complex_bb.py
@@ -7,7 +7,7 @@ import sdr
 def test_exceptions():
     with pytest.raises(ValueError):
         # x_r must be 1D
-        x_r = np.zeros((2, 2), dtype=np.float32)
+        x_r = np.zeros((2, 2), dtype=float)
         sdr.to_complex_bb(x_r)
     with pytest.raises(ValueError):
         # x_r must be real
