+++ tests/simulation/test_awgn.py
@@ -13,7 +13,7 @@ def test_exceptions():
 def test_real_from_snr():
     A = np.random.uniform(2, 3)  # Signal amplitude
     N = 1000
-    x = A * np.ones(N, dtype=np.float32)  # Use a constant signal so the power measurement has no uncertainty
+    x = A * np.ones(N, dtype=float)  # Use a constant signal so the power measurement has no uncertainty
 
     snr = np.random.uniform(1, 2)  # dB
     snr_linear = 10 ** (snr / 10)
@@ -47,7 +47,7 @@ def test_complex_from_snr():
 def test_real_from_noise():
     A = np.random.uniform(2, 3)  # Signal amplitude
     N = 1000
-    x = A * np.ones(N, dtype=np.float32)  # Use a constant signal so the power measurement has no uncertainty
+    x = A * np.ones(N, dtype=float)  # Use a constant signal so the power measurement has no uncertainty
 
     noise_std = np.random.uniform(1, 2)
     y = sdr.awgn(x, noise=noise_std**2)
